+++ hikka/utils.py
@@ -657,6 +657,7 @@ def get_named_platform() -> str:
     is_okteto = "OKTETO" in os.environ
     is_docker = "DOCKER" in os.environ
     is_heroku = "DYNO" in os.environ
+    is_codespaces = "CODESPACES" in os.environ
 
     if is_heroku:
         return "♓️ Heroku"
@@ -669,9 +670,34 @@ def get_named_platform() -> str:
 
     if is_okteto:
         return "☁️ Okteto"
+    
+    if is_codespaces:
+        return "🐈‍⬛ Codespaces"
 
     is_lavhost = "LAVHOST" in os.environ
     return f"✌️ lavHost {os.environ['LAVHOST']}" if is_lavhost else "📻 VDS"
+
+
+def get_platform_emoji() -> str:
+    BASE = (
+        '<emoji document_id="{}">🌘</emoji><emoji'
+        ' document_id="5195311729663286630">🌘</emoji><emoji'
+        ' document_id="5195045669324201904">🌘</emoji>'
+    )
+
+    if "OKTETO" in os.environ:
+        return BASE.format(5192767786174128165)
+
+    if "CODESPACES" in os.environ:
+        return BASE.format(5194976881127989720)
+
+    if "DYNO" in os.environ:
+        return BASE.format(5192845434887873156)
+
+    if "com.termux" in os.environ.get("PREFIX", ""):
+        return BASE.format(5193051778001673828)
+
+    return BASE.format(5192765204898783881)
 
 
 def uptime() -> int: